common: move function prototypes to their own header
common to menu.h, advanced.h, umdvideo_list.h
